third_party/WebKit/LayoutTests/compositing/squashing/do-not-squash-scroll-child-with-composited-descendants.html

<!DOCTYPE HTML>
<style>
#scroller {
  width: 200px;
  height: 200px;
  border: 1px solid black;
  overflow: scroll;
}

.scrolled {
  width: 100px;
  height: 100px;
  background: papayawhip;
  border: 1px solid black;
  margin: 10px;
  position: relative;
}

#fixed {
  width: 80px;
  height: 80px;
  background: blue;
  will-change: transform;
  position: fixed;
  top: 0px;
  left: 0px;
}

#composited {
  width: 20px;
  height: 20px;
  background: green;
  position: relative;
  will-change: transform;
}

#unsquashed {
  z-index: 1;
}
</style>
<div id="scroller">
  <div id="fixed"></div>
  <div class="scrolled"></div>
  <div class="scrolled"></div>
  <div id="unsquashed" class="scrolled">
    <div id="composited"></div>
  </div>
</div>

<script src='../../resources/js-test.js'></script>


<script>
description('Verifies that scroll children with composited descendants' +
            ' do not squash.');

if (window.internals)
  internals.settings.setPreferCompositingToLCDTextEnabled(true);

function scrollChildDoesNotSquash(layers) {
  var foundParent = false;
  var foundChild = false;
  layers["layers"].forEach(function(layer) {
    // If the child's bounds are 20x20, we're the scroll child with the
    // composited descendant. We should not have squashed and therefore our
    // bounds should remain 102x102 (100x100 plus the 1px border).
    if (layer.bounds &&  layer.bounds[0] == 20 && layer.bounds[1] == 20)
      foundChild = true;
    if (layer.bounds && layer.bounds[0] == 102 && layer.bounds[1] == 102)
      foundParent = true;
  });
  return foundParent && foundChild;
}

onload = function() {
  if (!window.internals)
    return;
  layers = JSON.parse(internals.layerTreeAsText(document));
  shouldBe('scrollChildDoesNotSquash(layers)', 'true');
};

</script>
