third_party/WebKit/LayoutTests/compositing/squashing/do-not-squash-scroll-child-with-composited-descendants-expected.txt

Verifies that scroll children with composited descendants do not squash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS scrollChildDoesNotSquash(layers) is true
